requirements.txt
flask
requests
gunicorn
python-whois
huggingface-hub
python-dotenv
